refactor(collections): move collection classes to the core.collections..
...package
