+++ core/com.b2international.snowowl.core/META-INF/MANIFEST.MF
@@ -38,6 +38,7 @@ Export-Package: com.b2international.snowowl.core,
  com.b2international.snowowl.core.bundle,
  com.b2international.snowowl.core.client,
  com.b2international.snowowl.core.codesystem,
+ com.b2international.snowowl.core.collection,
  com.b2international.snowowl.core.commit,
  com.b2international.snowowl.core.compare,
  com.b2international.snowowl.core.conceptmap,
@@ -72,7 +73,6 @@ Export-Package: com.b2international.snowowl.core,
  com.b2international.snowowl.core.rate,
  com.b2international.snowowl.core.repository,
  com.b2international.snowowl.core.request,
- com.b2international.snowowl.core.request.collection,
  com.b2international.snowowl.core.request.ecl,
  com.b2international.snowowl.core.request.expand,
  com.b2international.snowowl.core.request.io,
