+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/collection/TerminologyResourceCollectionCreateRequest.java
@@ -13,11 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.request.collection;
+package com.b2international.snowowl.core.collection;
+
+import java.util.TreeSet;
+import java.util.stream.Collectors;
 
 import org.hibernate.validator.constraints.NotEmpty;
 
-import com.b2international.snowowl.core.TerminologyResourceCollection;
+import com.b2international.commons.exceptions.BadRequestException;
+import com.b2international.snowowl.core.ResourceTypeConverter;
+import com.b2international.snowowl.core.domain.TransactionContext;
 import com.b2international.snowowl.core.internal.ResourceDocument.Builder;
 import com.b2international.snowowl.core.request.resource.BaseTerminologyResourceCreateRequest;
 
@@ -45,5 +50,22 @@ final class TerminologyResourceCollectionCreateRequest extends BaseTerminologyRe
 		return super.completeResource(builder)
 				.childResourceType(childResourceType);
 	}
-
+	
+//	@Override
+//	protected void preExecute(TransactionContext context) {
+//		var supportedChildResourceTypes = context.service(ResourceTypeConverter.Registry.class).getResourceTypeConverters()
+//				.values()
+//				.stream()
+//				.filter(ResourceTypeConverter::canBeContainedByCollection)
+//				.map(ResourceTypeConverter::getResourceType)
+//				.collect(Collectors.toCollection(TreeSet::new));
+//		
+//		if (!supportedChildResourceTypes.contains(childResourceType)) {
+//			throw new BadRequestException("'%s' is not supported as child resource type in collections. Select one from the following supported resource types: '%s'", childResourceType, supportedChildResourceTypes);
+//		}
+//		
+//		// FIXME call preExecute after validating child resource type as the method has a side effect on creating a new branch, which is not rolled back in case of an error 
+//		super.preExecute(context);
+//	}
+	
 }
