+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/collection/TerminologyResourceCollectionGetRequest.java
@@ -13,11 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.request.collection;
+package com.b2international.snowowl.core.collection;
 
 import com.b2international.snowowl.core.ResourceURI;
-import com.b2international.snowowl.core.TerminologyResourceCollection;
-import com.b2international.snowowl.core.TerminologyResourceCollections;
 import com.b2international.snowowl.core.request.resource.BaseGetResourceRequest;
 
 /**
