+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/collection/TerminologyResourceCollectionSearchRequest.java
@@ -13,12 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.request.collection;
+package com.b2international.snowowl.core.collection;
 
 import com.b2international.index.Hits;
 import com.b2international.index.query.Expressions.ExpressionBuilder;
-import com.b2international.snowowl.core.TerminologyResourceCollection;
-import com.b2international.snowowl.core.TerminologyResourceCollections;
 import com.b2international.snowowl.core.domain.RepositoryContext;
 import com.b2international.snowowl.core.internal.ResourceDocument;
 import com.b2international.snowowl.core.request.resource.BaseTerminologyResourceSearchRequest;
