+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/collection/TerminologyResourceCollectionSearchRequestBuilder.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.request.collection;
+package com.b2international.snowowl.core.collection;
 
-import com.b2international.snowowl.core.TerminologyResourceCollections;
+import com.b2international.snowowl.core.collection.TerminologyResourceCollectionSearchRequest.OptionKey;
 import com.b2international.snowowl.core.domain.RepositoryContext;
 import com.b2international.snowowl.core.request.SearchResourceRequest;
-import com.b2international.snowowl.core.request.collection.TerminologyResourceCollectionSearchRequest.OptionKey;
 import com.b2international.snowowl.core.request.resource.BaseTerminologyResourceSearchRequestBuilder;
 
 /**
